+++ artemis-weaver/src/main/java/com/artemis/weaver/CommonClassWeaver.java
@@ -3,12 +3,16 @@ package com.artemis.weaver;
 
 import static com.artemis.meta.ClassMetadataUtil.superName;
 
+import java.util.Arrays;
+import java.util.List;
+
 import org.objectweb.asm.AnnotationVisitor;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.Opcodes;
 
 import com.artemis.Weaver;
 import com.artemis.meta.ClassMetadata;
+import com.artemis.meta.ClassMetadata.WeaverType;
 
 class CommonClassWeaver extends ClassVisitor implements Opcodes {
 
@@ -21,6 +25,13 @@ class CommonClassWeaver extends ClassVisitor implements Opcodes {
 	
 	@Override
 	public void visit(int version, int access, String name, String signature, String superName, String[] interfaces) {
+		if (meta.annotation == WeaverType.PACKED) {
+			List<String> interfaceList = Arrays.asList("com/artemis/PackedComponent$DisposedWithWorld");
+			if (interfaces != null) {
+				interfaceList.addAll(Arrays.asList(interfaces));
+			}
+			interfaces = interfaceList.toArray(new String[0]);
+		}
 		cv.visit(version, access, name, signature, superName(meta), interfaces);
 	}