+++ artemis/src/main/java/com/artemis/ComponentTypeFactory.java
@@ -1,4 +1,4 @@
-package com.artemis;
+package com.artemis; 
 
 import java.util.IdentityHashMap;